Phase 6: Add TelegramLogRepository tests and fix compilation errors

+++ app/src/main/java/com/chris/m3usuite/telegram/core/T_TelegramFileDownloader.kt
@@ -192,7 +192,7 @@ class T_TelegramFileDownloader(
                         TelegramLogRepository.logFileDownload(
                             fileId = fileId,
                             progress = 0,
-                            total = result.result.expectedSize ?: 0,
+                            total = (result.result.expectedSize ?: 0).toInt(),
                             status = "started"
                         )
                         true
